requirements.txt
scipy==1.5.2
pandas==1.1.5
seaborn==0.11.1
numpy==1.17.0
matplotlib==3.3.4
numpy==1.21.2
Pillow==8.3.1
